@rafaqz 2023-04-21
Looks like we need more tests on `write` from other geometry types. I might implement `GeoInterface.convert` for `Feature` and `FeatureCollection` to make that easier to do first.