@visr 2023-04-22
Looks good, nice speedup and fix! The extra converts would be good, in a separate PR I assume.